ChemicalDataAssimilation

Documentation for ChemicalDataAssimilation.

ChemicalDataAssimilation.parse_rxns

ChemicalDataAssimilation.read_fac_file

ChemicalDataAssimilation.species_and_stoich

ChemicalDataAssimilation.parse_rxns — Method

function parse_rxns(rxns)

Given a list of equations rxns, parse equation returning a list of unique species in the entire mechanism and a list of reaction components, i.e. tuples of (reactants, reactants_stoich, products, products_stoich, rate_equation)

ChemicalDataAssimilation.read_fac_file — Method

function read_fac_file(path::String)

Parse a mechanism file in the FACSIMILE format (ending with .fac). This assumes that the file has all necessary sections to define the master chemical mechanism including

VOCs: a list of VOCs tracked in the simulation

generic_rate_coefficients

complex_rate_coefficients

peroxy_radicals: the species being summed to form RO2

reaction_definitions: The list of chemical reactions in our mechanism

Output

Returns a dictionary containing the above sections as text.

ChemicalDataAssimilation.species_and_stoich — Method

function species_and_stoich(rxn_half)

Given half of a reaction (either reactants or products), parse the equation and return a list of species involved and associated stoichiometric coefficients. If no species is involved (production or destruction), set the species to nothing.